Questão 05

+++ questao5.c
@@ -2,15 +2,15 @@
 
 int main(void)
 {
-    int contaDigitos = 0; 
-    int numero;
+    unsigned long contaDigitos = 0; 
+    unsigned long numero;
     printf("Digite um Número: ");
-    scanf("%d", &numero);
+    scanf("%ld", &numero);
     do
     {
         contaDigitos = contaDigitos + 1;
         numero = numero / 10;
     } while (numero != 0);
-    printf("Número de Dígitos: %d\n", contaDigitos);
+    printf("Número de Dígitos: %ld\n", contaDigitos);
     return 0;
 }
